% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abdatool-package.R
\docType{package}
\name{abdatool-package}
\alias{abdatool}
\alias{abdatool-package}
\title{abdatool: ABDA Tool}
\description{
Shiny App that shows voting results for the City of Zurich.
}
\author{
\strong{Maintainer}: Aline Metzler \email{aline.metzler@zuerich.ch}

}
\keyword{internal}
